requirements.txt
streamlit==1.18.0
youtube_transcript_api==0.6.0
punctuator==0.9.6
langchain==0.0.161
regex==2023.5.5
requests==2.30.0
